Text issue update
Updated issue where acceleration was not accurately displaying based on force.

The division by 10 and incomplete dealings with the currentMass variable led to the numbers being off in different ways on different levels. Since the mass cannot be changed and is always at 1, acceleration will always be equal to the force (F=ma). Current simple fix is to set acceleration equal to the current force